requirements.txt
asgiref==3.2.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1
django-appconf==1.0.4
django-countries==7.1
django-extensions==3.1.3
django-imagekit==4.0.2
django-multiupload==0.5.2
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
pilkit==2.0
Pillow==8.2.0
psycopg2==2.8.6
pycparser==2.20
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
